+++ app/src/main/java/com/todoroo/astrid/activity/TaskEditFragment.kt
@@ -324,6 +324,7 @@ class TaskEditFragment : Fragment() {
                 }
             },
             requestFocus = requestFocus,
+            multiline = viewState.multilineTitle,
         )
     }